requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
allure-pytest
lxml
mysql-connector-python-rf




